backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-RESTX==1.3.2
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pandas==2.0.3
numpy==1.24.3

# Celery and Async Task Queue
celery==5.3.4
redis==5.0.1

# Task Result Backend
flower==2.0.1
importlib-metadata==6.11.0

# Testing Dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Celery and Async Task Queue
celery==5.3.4
redis==5.0.1

# Task Result Backend
flower==2.0.1
